Fixed cpu

+++ chippy-8/Emulator/Interfaces/IDisplayDevice.cs
@@ -4,10 +4,10 @@ namespace chippy8
 {
 	public interface IDisplayDevice : IComponent
 	{
-		byte this [ushort i] { get; set; }
+		byte this [int i] { get; set; }
 		void Clear ();
-		bool CheckPixel (ushort pos);
-		void SetPixel (ushort pos);
+		bool CheckPixel (int pos);
+		void SetPixel (int pos);
 		void Update ();
 		void Draw ();
 	}
